pyproject.toml
[project]
name = "ai_chatbot"
authors = [
    { name = "Sanjay Kumar", email = "sanjay.kumar001@gmail.com"}
]
description = "Intelligent AI Chatbot with ERPNext Integration"
requires-python = ">=3.14"
readme = "README.md"
dynamic = ["version"]
dependencies = [
     "twilio==8.5.0",
]

[project.optional-dependencies]
idp = [
     "pypdf>=4.0",
     "openpyxl>=3.1",
     "python-docx>=1.1",
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"

[tool.ruff]
line-length = 110
target-version = "py314"

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "I",
    "UP",
    "B",
    "RUF",
]
ignore = [
    "B017", # assertRaises(Exception) - should be more specific
    "B018", # useless expression, not assigned to anything
    "B023", # function doesn't bind loop variable - will have last iteration's value
    "B904", # raise inside except without from
    "E101", # indentation contains mixed spaces and tabs
    "E402", # module level import not at top of file
    "E501", # line too long
    "E741", # ambiguous variable name
    "F401", # "unused" imports
    "F403", # can't detect undefined names from * import
    "F405", # can't detect undefined names from * import
    "F722", # syntax error in forward type annotation
    "W191", # indentation contains tabs
    "UP030", # Use implicit references for positional format fields (translations)
    "UP031", # Use format specifiers instead of percent format
    "UP032", # Use f-string instead of `format` call (translations)
    "UP037", # quoted annotations
    "UP040", # Use type aliases instead of type annotations
]
typing-modules = ["frappe.types.DF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true
